Create forensis.py
This code opens a disk image and a file system on the disk using the Img_Info and FS_Info classes, respectively. It then checks the file system type and, if it is NTFS, iterates through the root directory and prints the name and size of each file. The code also demonstrates how to extract deleted files from the file system and how to upload the disk image to cloud storage using the cloudstorage library.